implemented recomputing similarity score using cosine similarity, implemented saving filters with mapped codes, modified SOLR schema and config
