+++ requirements.txt
@@ -21,3 +21,6 @@ pysolr
 flask-socketio
 gevent-websocket
 eventlet
+fuzzywuzzy
+scikit-learn
+nltk
